Continue moving functions into their own translation units
