+++ commands/lfg.py
@@ -1,0 +1,2 @@
+async def createLFGPost(message):
+    print("yo")
